given that there's already a MultipleReporter and when you actually use the ReporterDecorator to filter the test reports in arbitrary ways, it's way more sensible to split up display and accounting taks in there: hence the pass/fail/error counting is now done through separate methods, while the paintXXX() mathods actually do what they advertise: only data DISPLAY. This means you can now actually derive Reporter classes which override the paintFail and/or paintPass methods, where the parent methods are then not invoked, and still have your pass/fail/etc. counts in order.
Also extended the MultiReporter and ReporterDecorator to present all methods available in the SimpleReporter and scorer classes.

For backwards compatibility, the 'preferred' reporter list has been augmented so that you still have a default 'no pass message display' behaviour while the SimpleReporter and its Html and Text derivatives now paint the 'pass' items too: this is accomplished by using the already present decorator base class and deriving a NoPassesReporter class from it, which is chained with the Html/TextReporter.
All this means that technically, the decorators can finally be used as meant: to manipulate any reporter/scorer method's behaviour.